fix: update sponsor routes to use correct database clients
- Update sponsor upload to use admin client
- Add public sponsor fetch endpoint
- Fix permission errors in preview
