+++ utils/database.ts
@@ -38,38 +38,4 @@ export async function getSponsors() {
   return data as (Sponsor & { sponsor_level: SponsorLevel })[];
 }
 
-export async function createSponsor(sponsor: Omit<Sponsor, 'id' | 'created_at' | 'updated_at'>) {
-  try {
-    console.log('Creating sponsor with data:', sponsor);
-    
-    const { data, error } = await supabaseAdmin
-      .from('sponsors')
-      .insert(sponsor)
-      .select(`
-        *,
-        sponsor_level:sponsor_levels(*)
-      `)
-      .single();
-    
-    if (error) {
-      console.error('Supabase error creating sponsor:', {
-        error,
-        errorMessage: error.message,
-        errorDetails: error.details,
-        errorHint: error.hint,
-        sponsor,
-      });
-      throw new Error(`Failed to create sponsor: ${error.message}${error.hint ? ` (${error.hint})` : ''}`);
-    }
-    
-    console.log('Successfully created sponsor:', data);
-    return data as Sponsor & { sponsor_level: SponsorLevel };
-  } catch (error) {
-    console.error('Error in createSponsor:', {
-      error,
-      sponsor,
-      errorMessage: error instanceof Error ? error.message : 'Unknown error',
-    });
-    throw error;
-  }
-}
+
