+++ utils/supabase-admin.ts
@@ -1,4 +1,5 @@
 import { createClient } from '@supabase/supabase-js';
+import type { Sponsor, SponsorLevel } from '../types/database';
 
 if (!process.env.NEXT_PUBLIC_SUPABASE_URL) {
   throw new Error('Missing env.NEXT_PUBLIC_SUPABASE_URL');
@@ -21,3 +22,39 @@ export const supabaseAdmin = createClient(
     },
   }
 );
+
+export async function createSponsor(sponsor: Omit<Sponsor, 'id' | 'created_at' | 'updated_at'>) {
+  try {
+    console.log('Creating sponsor with data:', sponsor);
+    
+    const { data, error } = await supabaseAdmin
+      .from('sponsors')
+      .insert(sponsor)
+      .select(`
+        *,
+        sponsor_level:sponsor_levels(*)
+      `)
+      .single();
+    
+    if (error) {
+      console.error('Supabase error creating sponsor:', {
+        error,
+        errorMessage: error.message,
+        errorDetails: error.details,
+        errorHint: error.hint,
+        sponsor,
+      });
+      throw new Error(`Failed to create sponsor: ${error.message}${error.hint ? ` (${error.hint})` : ''}`);
+    }
+    
+    console.log('Successfully created sponsor:', data);
+    return data as Sponsor & { sponsor_level: SponsorLevel };
+  } catch (error) {
+    console.error('Error in createSponsor:', {
+      error,
+      sponsor,
+      errorMessage: error instanceof Error ? error.message : 'Unknown error',
+    });
+    throw error;
+  }
+}
